+++ CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes will be documented here.
 ## [1.0.0] - 2025-02-07
 
 ### Changed
-- **Optimized** the logic of **list item composition** in `ListWC.java`.
+- **Optimized** the logic of list items initialization in `ListWC.java`.
 - **Improved** retry logic for `UiElement.java `**assertions**.
-- Updated allurium steps presets
+- **Updated** ListWC "should" assertion me[thod. Now it logs all conditions (arguments) into allure step.
+- **Updated** allurium steps presets
+- **Updated** steps description in `UiSteps.java`
 - Made Selenide timeout configurable based on **Allurium** `retryAmount` and `retryIntervalMs` parameters.
-    - The configuration timeout is now i**nitially set to `retryAmount * retryIntervalMs`.
+    - The configuration tim`]()`eout ]()is now i**nitially set to `retryAmount * retryIntervalMs`.